% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_functions.R
\name{apps.patch}
\alias{apps.patch}
\title{Updates the specified Application resource. You can update the following fields: auth_domain (https://cloud.google.com/appengine/docs/admin-api/reference/rest/v1/apps#Application.FIELDS.auth_domain) default_cookie_expiration (https://cloud.google.com/appengine/docs/admin-api/reference/rest/v1/apps#Application.FIELDS.default_cookie_expiration)}
\usage{
apps.patch(Application, appsId, updateMask = NULL)
}
\arguments{
\item{Application}{The \link{Application} object to pass to this method}

\item{appsId}{Part of `name`}

\item{updateMask}{Standard field mask for the set of fields to be updated}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/appengine/docs/admin-api/}{Google Documentation}

Other Application functions: \code{\link{Application}},
  \code{\link{apps.create}}
}
