% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_functions.R
\name{apps.repair}
\alias{apps.repair}
\title{Recreates the required App Engine features for the application in your project, for example a Cloud Storage bucket or App Engine service account. Use this method if you receive an error message about a missing feature, for example '*Error retrieving the App Engine service account*'.}
\usage{
apps.repair(RepairApplicationRequest, appsId)
}
\arguments{
\item{RepairApplicationRequest}{The \link{RepairApplicationRequest} object to pass to this method}

\item{appsId}{Part of `name`}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/appengine/docs/admin-api/}{Google Documentation}

Other RepairApplicationRequest functions: \code{\link{RepairApplicationRequest}}
}

