% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{CpuUtilization}
\alias{CpuUtilization}
\title{CpuUtilization Object}
\usage{
CpuUtilization(aggregationWindowLength = NULL, targetUtilization = NULL)
}
\arguments{
\item{aggregationWindowLength}{Period of time over which CPU utilization is calculated}

\item{targetUtilization}{Target CPU utilization ratio to maintain when scaling}
}
\value{
CpuUtilization object
}
\description{
CpuUtilization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Target scaling by CPU usage.
}

