% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{DiskUtilization}
\alias{DiskUtilization}
\title{DiskUtilization Object}
\usage{
DiskUtilization(targetWriteBytesPerSecond = NULL,
  targetWriteOpsPerSecond = NULL, targetReadBytesPerSecond = NULL,
  targetReadOpsPerSecond = NULL)
}
\arguments{
\item{targetWriteBytesPerSecond}{Target bytes written per second}

\item{targetWriteOpsPerSecond}{Target ops written per second}

\item{targetReadBytesPerSecond}{Target bytes read per second}

\item{targetReadOpsPerSecond}{Target ops read per seconds}
}
\value{
DiskUtilization object
}
\description{
DiskUtilization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Target scaling by disk usage. Only applicable for VM runtimes.
}
