% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{NetworkUtilization}
\alias{NetworkUtilization}
\title{NetworkUtilization Object}
\usage{
NetworkUtilization(targetSentBytesPerSecond = NULL,
  targetSentPacketsPerSecond = NULL, targetReceivedBytesPerSecond = NULL,
  targetReceivedPacketsPerSecond = NULL)
}
\arguments{
\item{targetSentBytesPerSecond}{Target bytes sent per second}

\item{targetSentPacketsPerSecond}{Target packets sent per second}

\item{targetReceivedBytesPerSecond}{Target bytes received per second}

\item{targetReceivedPacketsPerSecond}{Target packets received per second}
}
\value{
NetworkUtilization object
}
\description{
NetworkUtilization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Target scaling by network usage. Only applicable for VM runtimes.
}

