% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{Operation}
\alias{Operation}
\title{Google App Engine Admin API Objects 
Provisions and manages App Engine applications.}
\usage{
Operation(Operation.metadata = NULL, Operation.response = NULL,
  name = NULL, metadata = NULL, done = NULL, error = NULL,
  response = NULL)
}
\arguments{
\item{Operation.metadata}{The \link{Operation.metadata} object or list of objects}

\item{Operation.response}{The \link{Operation.response} object or list of objects}

\item{name}{The server-assigned name, which is only unique within the same service that originally returns it}

\item{metadata}{Service-specific metadata associated with the operation}

\item{done}{If the value is `false`, it means the operation is still in progress}

\item{error}{The error result of the operation in case of failure}

\item{response}{The normal response of the operation in case of success}
}
\value{
Operation object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:15:40
filename: /Users/mark/dev/R/autoGoogleAPI/googleappenginev1.auto/R/appengine_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Operation Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This resource represents a long-running operation that is the result of a network API call.
}
\seealso{
Other Operation functions: \code{\link{Operation.metadata}},
  \code{\link{Operation.response}}
}

