% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{OperationMetadataV1}
\alias{OperationMetadataV1}
\title{OperationMetadataV1 Object}
\usage{
OperationMetadataV1(method = NULL, insertTime = NULL, endTime = NULL,
  user = NULL, target = NULL)
}
\arguments{
\item{method}{API method that initiated this operation}

\item{insertTime}{Time that this operation was created}

\item{endTime}{Time that this operation completed}

\item{user}{User who requested this operation}

\item{target}{Name of the resource that this operation is acting on}
}
\value{
OperationMetadataV1 object
}
\description{
OperationMetadataV1 Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata for the given google.longrunning.Operation.
}

